+++ src/primitives/line.js
@@ -11,7 +11,7 @@ import { getBasicMaterial } from '../shader.js';
 // for the high-level description of what is being rendered.
 // Differently from WL's Line, our lines aren't affected by
 // lightning and therefore don't have VertexNormals.
-export default function ({ color, coords, opacity = 1 }, extent) {
+export default function ({ color = [0, 0, 0], coords, opacity = 1 }, extent) {
 	return new Line(
 		new BufferGeometry().setAttribute(
 			'position',