+++ CHANGES.md
@@ -14,6 +14,7 @@ Improvements:
 - Improve performance of cones by ~6x (#64)
 - Improve performance of arrows (#66)
 - Improve performance of cylinders (#67)
+- Add default colors to primitives and lights (#75)
 
 Internals:
 - Convert all materials to ShaderMaterial/RawShaderMaterial (#53)